Move database connection code to a submodule
