+++ libs/exo-sql/src/sql/mod.rs
@@ -18,9 +18,8 @@ mod test_util;
 
 pub mod array_util;
 pub mod column;
+pub mod connect;
 pub mod database;
-pub mod database_client;
-pub mod database_client_manager;
 pub mod function;
 pub mod order;
 pub mod physical_column;
